// apps/web/src/lib/api/generated/models/feedbackQuestion.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Formex
 * OpenAPI spec version: 0.0.1
 */

export interface FeedbackQuestion {
  id: string;
  institution_id: string;
  question: string;
  order_index: number;
  /** @nullable */
  created_at: string | null;
  /** @nullable */
  updated_at: string | null;
}